backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
motor>=3.5.0
pydantic[email]>=2.0.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
